+++ pkg/backupschedule/manager.go
@@ -1,5 +1,7 @@
 package backupschedule
 
+import "github.com/cloudogu/k8s-backup-operator/pkg/additionalimages"
+
 // maxTries controls the maximum number of waiting intervals between tries when getting an error that is recoverable
 // during k8s operations.
 var maxTries = 5
@@ -10,9 +12,9 @@ type defaultManager struct {
 	deleteManager
 }
 
-func NewManager(clientSet ecosystemInterface, recorder eventRecorder, namespace string, kubectlImage string) *defaultManager {
+func NewManager(clientSet ecosystemInterface, recorder eventRecorder, namespace string, imageConfig additionalimages.ImageConfig) *defaultManager {
 	return &defaultManager{
-		createManager: newCreateManager(clientSet, recorder, namespace, kubectlImage),
+		createManager: newCreateManager(clientSet, recorder, namespace, imageConfig),
 		updateManager: newUpdateManager(clientSet, recorder, namespace),
 		deleteManager: newDeleteManager(clientSet, recorder, namespace),
 	}