+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.9.0] - 2023-12-08
+### Added
+- [#14] Patch template for mirroring this component and its images
+### Changed
+- [#17] Replace create-backup-script with an operator subcommand.
+  This way, the backup schedule cron job can use the same image as the operator.
+- [#15] Delete kustomize structure and hold the operator yaml files just in a helm chart.
+
 ## [v0.8.0] - 2023-12-04
 ### Added
 - [#19] Sync backup list with provider on operator startup